+++ tests/testthat/test-server.R
@@ -22,7 +22,8 @@ test_that("validate pjnz", {
          errors = list(),
          data = list(hash = "12345",
                      type = "pjnz",
-                     data = list(country = "Botswana"),
+                     data = list(country = "Botswana",
+                                 iso3 = "BWA"),
                      filename = "original.PJNZ",
                      filters = NULL)))
 })
@@ -196,7 +197,7 @@ test_that("model interactions", {
 
 test_that("plotting metadata is exposed", {
   server <- hintr_server()
-  r <- httr::GET(paste0(server$url, "/meta/plotting/", "Malawi"))
+  r <- httr::GET(paste0(server$url, "/meta/plotting/", "MWI"))
   expect_equal(httr::status_code(r), 200)
   response <- response_from_json(r)
   expect_equal(response$status, "success")